requirements.txt
discord
python-dotenv
discord-pretty-help
wavelink==2.6.4
vt-py
tqdm
requests